third_party/WebKit/LayoutTests/printing/page-rule-specificity.html

<!DOCTYPE html>
<html>
<head>
    <title>Page rule specificity</title>
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
    <style>
        @page :first {
            margin-left: 100px;
        }
        @page {
            margin-left: 50px;
        }
    </style>
</head>
<body>
    <p id="log"></p>
    <script>
        test(function(){
            assert_true(window.testRunner instanceof Object);
            assert_true(window.internals instanceof Object);
        }, "Test needs testRunner framework.");

        test(function(){
            assert_equals(internals.pageProperty("margin-left", 0), "100", "Left margin.");
        }, ":first specificity beats order of appearance.");

    </script>
</body>
</html>
